Updates to the ContractBundle, including new "AbstractVariableModel" model that contains logic for general template variable usage.
This class will extend to meet the specific variable requirements for each contract in the generator templates.

This object is then passed to the Generator itself so it can use it to populate the specific contract template.

Updated the BandBundles "DependecyInjection" folder name (updated to "DependencyInjection". Fixed a service in this bundle to (form).